+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Fixed files with special characters not being recognized (ex. `./src/[...something].ts`)
+
 ## [2.0.1] - 2023-08-06
 
 - Fixed generics not being sorted